refactor: generify e2 special muls
